rename player_bar to player; move to widgets
