+++ lib/ui/widgets/player.dart
